requirements.txt
pyupbit
python-dotenv
pandas
requests
